Refactored env into the vec_env directory
