+++ rl_algo_impls/shared/vec_env/__init__.py
@@ -1,0 +1,1 @@
+from rl_algo_impls.shared.vec_env.make_env import make_env, make_eval_env
